removed react-sdk package
