--- packages/react-sdk/package.json
@@ -1,45 +1,0 @@
-{
-  "name": "@corbado/react-sdk",
-  "version": "2.1.2-alpha.0",
-  "description": "This SDK is for developers who want to add Corbado's Authentication management in their applications but want to use a UI flow of their own.",
-  "author": "Abdullah Shahbaz <abdullah_ghani@live.com>",
-  "homepage": "https://github.com/corbado/javascript#readme",
-  "license": "ISC",
-  "main": "dist/index.js",
-  "module": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "directories": {
-    "lib": "dist"
-  },
-  "files": [
-    "dist"
-  ],
-  "publishConfig": {
-    "access": "public"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/corbado/javascript.git"
-  },
-  "scripts": {
-    "clean": "rimraf ./dist",
-    "build": "webpack --config webpack.prod.js",
-    "build:dev": "npm run clean && tsc -b",
-    "lint:fix": "eslint src --ext .js,.jsx,.ts,.tsx --fix",
-    "lint": "eslint src --ext .js,.jsx,.ts,.tsx"
-  },
-  "bugs": {
-    "url": "https://github.com/corbado/javascript/issues"
-  },
-  "peerDependencies": {
-    "react": ">=18.0.0"
-  },
-  "dependencies": {
-    "@corbado/web-core": "^2.1.2-alpha.0"
-  },
-  "devDependencies": {
-    "@corbado/types": "^2.1.2-alpha.0",
-    "@types/react": "^18.2.0"
-  },
-  "gitHead": "bdc9da96c3a34574545c5d9b522464ca9d3eca57"
-}
