--- packages/react-sdk/src/hooks/index.ts
@@ -1,1 +1,0 @@
-export { useCorbado } from './useCorbado';
