--- packages/react-sdk/src/index.ts
@@ -1,25 +1,0 @@
-import {
-  InvalidEmailError,
-  InvalidFullnameError,
-  InvalidOtpInputError,
-  InvalidPasskeyError,
-  NonRecoverableError,
-  NoPasskeyAvailableError,
-  PasskeyChallengeCancelledError,
-  UnknownUserError,
-  UserAlreadyExistsError,
-} from '@corbado/web-core';
-
-export * from './contexts';
-export * from './hooks';
-export {
-  InvalidPasskeyError,
-  NoPasskeyAvailableError,
-  InvalidOtpInputError,
-  PasskeyChallengeCancelledError,
-  UnknownUserError,
-  UserAlreadyExistsError,
-  InvalidFullnameError,
-  InvalidEmailError,
-  NonRecoverableError,
-};
