--- packages/react-sdk/webpack.prod.js
@@ -1,16 +1,0 @@
-const { merge } = require('webpack-merge');
-const path = require('path');
-const common = require('../../webpack.common.js');
-
-module.exports = merge(common, {
-  entry: path.resolve(__dirname, 'src/index.ts'),
-  output: {
-    path: path.resolve(__dirname, 'dist'),
-  },
-  devtool: 'source-map',
-  mode: 'production',
-  externals: {
-    '@corbado/web-core': '@corbado/web-core',
-    react: 'react',
-  },
-});
