+++ packages/react/src/index.ts
@@ -1,8 +1,8 @@
-import { useCorbado } from '@corbado/react-sdk';
 import { CorbadoThemes } from '@corbado/shared-ui';
 import type { CorbadoAuthConfig } from '@corbado/types';
 
 import CorbadoProvider, { type CorbadoProviderProps } from './hocs/CorbadoProvider';
+import { useCorbado } from './hooks/useCorbado';
 import CorbadoAuth from './screens/core/CorbadoAuth';
 import Login from './screens/core/Login';
 import PasskeyList from './screens/core/PasskeyList';
