+++ packages/react/src/screens/core/PasskeyList.tsx
@@ -1,10 +1,10 @@
-import { useCorbado } from '@corbado/react-sdk';
 import type { PassKeyList } from '@corbado/types';
 import type { FC } from 'react';
 import React, { useCallback, useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 
 import { LoadingSpinner, PasskeyCreate, PasskeyListErrorBoundary, PasskeyListItem } from '../../components';
+import { useCorbado } from '../../hooks/useCorbado';
 
 const PasskeyList: FC = () => {
   const { getPasskeys, globalError, isAuthenticated } = useCorbado();
